refactor trpc router and plain client
